half

+++ whisper/model.py
@@ -27,7 +27,8 @@ class ModelDimensions:
 
 class LayerNorm(nn.LayerNorm):
     def forward(self, x: Tensor) -> Tensor:
-        return super().forward(x.float()).type(x.dtype)
+        # return super().forward(x.float()).type(x.dtype) sovits5.0
+        return super().forward(x).type(x.dtype)
 
 
 class Linear(nn.Linear):
